requirements.txt
streamlit
pypdf
chromadb
sentence-transformers
google-generativeai
python-dotenv
pandas
openpyxl
python-docx
